use preact component for likes
